Analog Interface v1
Outputs the X and Y location as an analog voltage difference.

Currently it is set to millimeters, but can be adjusted to centimeters.
If mm range is limited to 4m, cm range is limited to 40m (Trade-off).

Code might not be perfect due to requiring Linux to compile (needs testing).