More work on webclient html skeleton (#38)
* Basic webclient skeleton

I've added in a basic skeleton for now. further changes to be made

* Created .xml file

* Update index.html

* Delete index.xml

to Avoid duplications. Further xml files are to be tested locally

* reformatted code since VS kept playing out

* Added in footer and a new webpage for online manual

Added in a footer according to Michael's sketch to for info relating to web client. Additionally a second webpage was created for an online manual (draft)

* Added in grid 1-3 and info

Basically coded in the necessary grids and added in the information section

* Added in CSS

Added in a css folder including the styles file for temporary use

* Minor fixes

* Further improvements to basic skeleton

* Updated HTML