WIP -  http://localhost/anchors not displaying correctly
Tried to create a simple script that would GET all anchor positions and display them on the webpage when the 'Testing' button was clicked.

ATM, it just erases the default '{x,y,z}' text (my guess is that the returned data is blank due to either incorrect parameters or something wrong with how the REST API is working on localhost).

@raminouri hoping you can look into this and see what you can find.